# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation de l'API pour la gestion des utilisateurs et l'authentification.",
        "contact": {
            "email": "sahliabderazak530@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recherche-candidat": {
            "get": {
                "tags": [
                    "Candidat"
                ],
                "summary": "Rechercher un candidat",
                "operationId": "rechercheCandidat",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "description": "Nom du candidat",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "prenom",
                        "in": "query",
                        "description": "Prénom du candidat",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidats trouvés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nom": {
                                                "type": "string",
                                                "example": "Doe"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "example": "John"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/users/unarchive/{id}": {
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Désarchiver un utilisateur",
                "description": "Désarchiver un utilisateur par son ID",
                "operationId": "784eb26a4d2795543f9606761640ff2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur désarchivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur désarchivé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/archive/{id}": {
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Archiver un utilisateur",
                "description": "Archiver un utilisateur par son ID",
                "operationId": "c728c68e408ac6b3242aba9eeb46b314",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur archivé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur archivé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/user/info": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Récupérer les informations de l'utilisateur connecté",
                "description": "Retourne les informations de l'utilisateur actuellement authentifié.",
                "operationId": "8cc5b97acd353191d7b83864b3d96806",
                "responses": {
                    "200": {
                        "description": "Informations de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nom_societe": {
                                            "type": "string",
                                            "example": "ABC Corp"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://example.com/storage/images/user.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Non authentifié"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recherche-candidat-archive": {
            "get": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Rechercher des candidats archivés",
                "description": "Rechercher des candidats archivés en fonction du nom et du prénom.",
                "operationId": "7540bcefa93bf13af83334f4241484eb",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Dupont"
                        }
                    },
                    {
                        "name": "prenom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Jean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des candidats archivés trouvés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Candidat"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun candidat trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun candidat trouvé."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recruteurs-archives/recherche": {
            "get": {
                "tags": [
                    "Recruteurs"
                ],
                "summary": "Rechercher des recruteurs archivés",
                "description": "Rechercher des recruteurs archivés par lettre.",
                "operationId": "b6b7d250870f1a6e6f949ef0f4e4e28b",
                "parameters": [
                    {
                        "name": "letter",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des recruteurs archivés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/recherche-recruteur": {
            "get": {
                "tags": [
                    "Recruteurs"
                ],
                "summary": "Rechercher des recruteurs",
                "description": "Rechercher des recruteurs par lettre.",
                "operationId": "e5e7589a7b22e4b8815a35de63fd0fc8",
                "parameters": [
                    {
                        "name": "letter",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des recruteurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Candidat": {
                "required": [
                    "nom",
                    "prenom",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "pays": {
                        "type": "string",
                        "example": "France"
                    },
                    "ville": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "codePostal": {
                        "type": "string",
                        "example": "75001"
                    },
                    "tel": {
                        "type": "string",
                        "example": "0102030405"
                    },
                    "niveauEtude": {
                        "type": "string",
                        "example": "Bac+5"
                    },
                    "niveauExperience": {
                        "type": "string",
                        "example": "3 ans"
                    },
                    "cv": {
                        "type": "string",
                        "example": "https://example.com/cvs/cv.pdf"
                    },
                    "offre_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "archived": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "numTel": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "image": {
                        "type": "string",
                        "example": "profile.jpg"
                    },
                    "archived": {
                        "type": "boolean",
                        "example": false
                    },
                    "nom_societe": {
                        "type": "string",
                        "example": "Example Company"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "apropos": {
                        "type": "string",
                        "example": "About the company"
                    },
                    "lien_site_web": {
                        "type": "string",
                        "example": "https://example.com"
                    },
                    "fax": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "domaine_activite": {
                        "type": "string",
                        "example": "Technology"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}